pipeline-serving/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipeline-serving"
version = "0.0.1"
authors = [
  { name="Gao Hongnan", email="hongnangao@gmail.com" },
]
description = "Serving pipeline."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "gaohn-common-utils",
    "pydantic==2.0.3",
    "mlflow==2.4.1",
    "uvicorn==0.23.0",
    "fastapi==0.100.0",
    "scikit-learn==1.3.0",
    "psutil",
    "python-dotenv==0.21.1",
    "optuna==3.2.0",
]

[project.optional-dependencies]
dev = [
    "bandit==1.7.5",
    "black==23.3.0",
    "colorama==0.4.4",
    "coverage==7.0.5",
    "isort==5.12.0",
    "mypy==1.3.0",
    "pylint==2.17.4",
    "pytest==6.2.5",
    "pytest-cov==4.0.0",
    "toml==0.10.2",
]

[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)

[project.urls]
"Homepage" = "https://github.com/gao-hongnan/TheBareOps"
"Bug Tracker" = "https://github.com/gao-hongnan/TheBareOps/issues"

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39']
verbose = false
diff = false
color = false
include = '\.pyi?$'
exclude = '''
/(
    \.eggs                      # exclude a few common directories in the
  | \.git                       # root of the project
  | \.hg
  | \.mypy_cache
  | \.tox
  | (.*/)?venv_.*/.*
  | (.*/)?\.venv/.*
  | (.*/)?venv/.*
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
verbose = false
skip = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    "venv_*",
    "venv",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
]
skip_glob = [
    "**/venv_*",
    "venv",
    ".venv",
]

[tool.pylint]
    [tool.pylint.'MASTER']
    ignore-paths = [
        "(.*\\/)?venv_.*\\/.*",
        "(.*\\/)?\\.venv\\/.*",
        "(.*\\/)?venv\\/.*",
    ]
    ignore = [
        ".eggs",
        ".git",
        ".hg",
        ".mypy_cache",
        ".tox",
        "venv_*",
        "venv",
        ".venv",
        "_build",
        "buck-out",
        "build",
        "dist",
    ]
    [tool.pylint.'MESSAGES CONTROL']
    disable = [
        "redefined-outer-name",
        "missing-module-docstring",
        "missing-class-docstring",
        "missing-function-docstring",
        "fixme",
        "reimported",
        "ungrouped-imports",
        "invalid-name",
    ]

# https://bandit.readthedocs.io/en/latest/config.html
[tool.bandit]
exclude_dirs = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    "venv_*",
    "venv",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
]

# https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml-file
[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
disable_error_code = ['override', 'attr-defined']
disallow_any_generics = false

[tool.pytest.ini_options]
# add any pytest options here, for example:
# addopts = --verbose --capture=no
testpaths = "tests"
filterwarnings = "ignore::DeprecationWarning" # ignore deprecation warnings when running tests